suggested changes from before PR was merged
