+++ doc/tutorials/pennylane_advanced_usage.py
@@ -1,6 +1,4 @@
 r"""
-.. role:: html(raw)
-   :format: html
 
 .. _advanced_features:
